Revert "Cambios"
This reverts commit 3b7136c5439eb9759956107e381a898c9c7aef9f.

+++ NodeJs/src/Views/FertilizanteApp.js
@@ -9,10 +9,21 @@ export default function FertilizanteApp() {
   const tableRef = useRef(null);
   const [selectdNameFertilizante, setSelectdNameFertilizante] = useState(null);
 
+  //FORMATO EN TABLA 
+  const capitalize = (str) => {
+    return str.replace(/\w\S*/g, (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
+  };
+
   const listarFertilizante = () => {
     FertilizanteService.findAll()
       .then((response) => {
-        setFertilizante(response.data);
+        const formatoData = response.data.map(item => ({
+          ...item,
+          domicilio_Distribuidora: capitalize(item.domicilio_Distribuidora),
+          marca: capitalize(item.marca)
+        }));
+        setFertilizante(formatoData);
+        console.log(formatoData);
       })
       .catch((error) => {
         console.log(error);
@@ -24,13 +35,17 @@ export default function FertilizanteApp() {
   }, []);
 
   const deleteFertilizante = (id) => {
-    FertilizanteService.delete(id)
-      .then((response) => {
-        listarFertilizante();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
+    const confirmarDelete = window.confirm("¿Estás seguro de que deseas eliminar este registro?");
+    if (confirmarDelete) {
+      FertilizanteService.delete(id)
+        .then((response) => {
+          listarFertilizante();
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    }
+
   };
 
   const handleRowClick = (id, name) => {
@@ -62,7 +77,7 @@ export default function FertilizanteApp() {
         <p>
           {selectdNameFertilizante
             ? `Nombre del Fertilizante: ${selectdNameFertilizante}`
-            : "No se esta seleccionando un Fertilizante"}
+            : ""}
         </p>
         <div className="table-container" ref={tableRef}>
           <div className="table-responsive">
@@ -76,6 +91,7 @@ export default function FertilizanteApp() {
                   <th>Costo unitario</th>
                   <th>Domicilio distribuidora</th>
                   <th>Fecha de caducidad</th>
+                  <th>Fecha de adquisicion</th>
                   <th>Lote</th>
                   <th>Marca</th>
                 </tr>
@@ -103,6 +119,7 @@ export default function FertilizanteApp() {
                     <td>{fertilizante.costo_Unitario}</td>
                     <td>{fertilizante.domicilio_Distribuidora}</td>
                     <td>{fertilizante.fecha_Caducidad}</td>
+                    <td>{fertilizante.fecha_Adquisicion}</td>
                     <td>{fertilizante.lote}</td>
                     <td>{fertilizante.marca}</td>
                   </tr>
@@ -129,18 +146,18 @@ export default function FertilizanteApp() {
                   Insertar
                 </button>
               </Link>
-              <Link to="/form-fertilizante">
+              <Link to="/">
                 <button
                   type="button"
                   className="btn btn-success"
                   class="btnimagen"
                 >
                   <img
-                    src="icons/Buscar.png"
-                    alt="Buscar fertilizante"
+                    src="icons/Regresar.png"
+                    alt="Regresar"
                     className="imgBuscar"
                   ></img>
-                  Consultar
+                  Regresar
                 </button>
               </Link>
             </>
